Make the link with Hitscore, Lwt, and PGOcaml
Now the webserver pretends to use hitscore, lwt.syntax, pgocaml.syntax
and their respective dependencies.
Some of the changes in Hitscore_std are just for testing purposes,
they will be removed. The change in app/hitscore.ml is due to a small
difference in the syntax parsed by ocaml and camlp4. Now the _tags
file is not completely autogenerated by OASIS (the last line is a
customization).