Delay the initialization of services (for static)
A recent version upgrade of Ocsigen seems to require the registration
AND the initialization of services to wait for the configuration file
to be parsed, hence this commit uses hidden global option references
instead of global values. The module Services also get two new
functions 'link' and 'register' to clean its API:
  Services.(register persons) (fun (transpose, highlight) () ->
    Display_service.make ~hsc:hitscore_configuration
      ~main_title:"People"
      (persons ~highlight ?transpose hitscore_configuration));
for registering the persons service, and
  Services.(link flowcell) [pcdata fcid] fcid;
to create a link (Eliom_output.Html5.a).