Start implementing table sorting (WIP)
A lot of hacking has been done to get the data and try to use it, but
this seems stuck in an Eliom bug (?). Indeed, the on_click handlers do
not seem to correspond to the ones I set (eprintf and the debug
service disagree). This should be investigated more.