Make Highcharts live happily with Eliom's onload
This patch manages to get rid of jQuery's
  $(document).ready(...)
which was not working properly with Eliom's application
infrastructure. For instance, when clicking on a link to the same
service, the charts were disappearing and the page needed to be
reloaded to see them again. The code that loads the charts is now in a
  Eliom_services.onload {{ ... }}
thanks to Js.Unsafe.eval_string.