Implement password-based chained authentication
This patch gives semantics to the field 'password_hash' in the layout:
  Digest.(string password |! to_hex)
the same string can be obtained with
  echo -n "password" | md5sum
The current 'chain' tries to authenticate with login and primary
email, first with the password-hash and then with the PAM module.
This is somewhat vulnerable to timing attacks (for now, if the user
does not exist, we do not call PAM, so on could try to guess existing
logins from the processing time).