Allow full details on more than one library
For now the only way to get these details with to add the
'show=details' argument to the URL of the /libraries service. Showing
details on a dozen of libraries makes the web-browser quite slow,
trying to show the details of all the current libraries (522) make the
chromium eat all the memory and crash the page.